requirements.txt
requests==2.31.0
python-dotenv==1.0.1
stravalib==1.7
seaborn==0.13.2
matplotlib==3.8.4